refactor(node): extract validator subcommands to commands/validator.rs
Eight cmd_validator_* functions move out of main.rs:
  cmd_validator_add, _unjail, _force_unjail, _transfer_admin,
  _list, _remove, _toggle, _rename

Same pattern as #641 (wallet cmds): pure CLI handlers, no consensus
path touched. Underlying mutations stay in sentrix-core::authority
and sentrix-staking.

`get_db_path` widened from `fn` to `pub(crate) fn` so the new module
can read the configured chain DB path via the same SENTRIX_DATA_DIR-
aware helper. No other visibility changes.

CLI dispatch updated: `cmd_validator_*` → `commands::validator::cmd_validator_*`.

main.rs: 4,552 → 4,344 LOC (−208 = −4.6%).
Combined with #641: 4,780 → 4,344 (−436 = −9.1%).
